+++ app/src/androidTest/java/com/teamyg/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.tjyg
+package com.teamyg
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.tjyg", appContext.packageName)
+        assertEquals("com.teamyg", appContext.packageName)
     }
 }
