+++ app/src/main/AndroidManifest.xml
@@ -10,12 +10,12 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.Tjyg">
+        android:theme="@style/Theme.TeamYg">
         <activity
             android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name"
-            android:theme="@style/Theme.Tjyg">
+            android:theme="@style/Theme.TeamYg">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
