+++ app/src/main/java/com/teamyg/MainActivity.kt
@@ -1,18 +1,18 @@
-package com.tjyg
+package com.teamyg
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import com.tjgy.feature.sample.SampleScreen
-import com.tjyg.ui.theme.TjygTheme
+import com.teamyg.ui.theme.TeamYGTheme
+import com.teamyg.feature.sample.SampleScreen
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
         setContent {
-            TjygTheme {
+            TeamYGTheme() {
                 SampleScreen()
             }
         }
