+++ app/src/main/java/com/teamyg/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.tjyg.ui.theme
+package com.teamyg.ui.theme
 
 import android.app.Activity
 import android.os.Build
@@ -34,7 +34,7 @@ private val LightColorScheme = lightColorScheme(
 )
 
 @Composable
-fun TjygTheme(
+fun TeamYGTheme(
     darkTheme: Boolean = isSystemInDarkTheme(),
     // Dynamic color is available on Android 12+
     dynamicColor: Boolean = true,
