+++ app/src/main/java/com/teamyg/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.tjyg.ui.theme
+package com.teamyg.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
