+++ app/src/test/java/com/teamyg/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.tjyg
+package com.teamyg
 
 import org.junit.Test
