+++ domain/src/main/java/com/teamyg/domain/MyClass.kt
@@ -1,0 +1,4 @@
+package com.teamyg.domain
+
+class MyClass {
+}
