--- domain/src/main/java/com/tjyg/domain/MyClass.kt
@@ -1,4 +1,0 @@
-package com.tjyg.domain
-
-class MyClass {
-}
